Refactored: Extracted main method to a separate class TennisGameMain
